Move language over to dev settings for now
